Add unfinished id_duplicate scripts and associated files (#87)
* Modify run.sh and migrate scripts from iPython notebooks to Python scripts

* Add unfinished id_duplicates, wrapper, function, and test files

Matt updated this version and added new modules. Before we integrate it fully, I'd like to add/run tests on various edge cases to ensure that it catches them. Considering we want to identify duplicates and identified duplicates will be merged/deleted, I want to be extra certain it functions alright. I figure it'll leave us with less headaches if we ensure ahead of time that we're not getting any false positives/negatives.

Developing and running those tests will take additional time. Additionally, I can't run the script on my Windows OS machine--one of the modules can only be installed reliably on Linux and Mac OS, which may require discussion as to whether we want a module that doesn't work on Windows OS -- this won't stop our Github automation, which runs on a Linux OS. But it does slow down anyone who wants to test or develop it on a Windows Machine. Let's make sure we're fine with that.