Modify test_id_duplicates, add related test material, and setup pytest framework (#91)
* Modify run.sh and migrate scripts from iPython notebooks to Python scripts

* Add unfinished id_duplicates, wrapper, function, and test files

Matt updated this version and added new modules. Before we integrate it fully, I'd like to add/run tests on various edge cases to ensure that it catches them. Considering we want to identify duplicates and identified duplicates will be merged/deleted, I want to be extra certain it functions alright. I figure it'll leave us with less headaches if we ensure ahead of time that we're not getting any false positives/negatives.

Developing and running those tests will take additional time. Additionally, I can't run the script on my Windows OS machine--one of the modules can only be installed reliably on Linux and Mac OS, which may require discussion as to whether we want a module that doesn't work on Windows OS -- this won't stop our Github automation, which runs on a Linux OS. But it does slow down anyone who wants to test or develop it on a Windows Machine. Let's make sure we're fine with that.

* Modify test_id_duplicates, add related test material, and setup pytest framework.

Add test_id_duplicates_bloomfield and input.csv file. This can be considered a prototype for additional tests.

Added __init__.py to /tests and /data_prep_scripts folders. These are needed to tell python to import files within them as modules.

* Add remaining id_duplicates_functions test material

* Update requirements.txt

* Update requirements.txt

* Update pytest.yml

Change python version from 3.6 to 3.7

* Update requirements.txt

* Add additional test cases