Updated tests/test_id_duplicates.py
Tests are now done against a snapshot of data stored in
`tests/test-data/test_id_duplicates.csv`. Note that changing this file
will likely cause tests to fail as ids are referenced explicitly.

As of this commit, the first test `test_bloomfield` fails. As far as I
can tell, this is a problem with the deduplication process, and not the
test case itself.