Major revision to test functionality
Removed prior method of obtaining and evaluating test data, based on using historical copy of entire dataset. This method was time-consuming, as it ran id_duplicates over the entire dataset, and risked being out of date and difficult to correct.

In its place, I modified test and helper functions such that each test includes only the data relevant to that particular test, and runs id_duplicates for that data alone.

To do this, the tests needed to be modified to fit into the new changes. This was done.

Similarly, removed a number of the helper functions, as they were associated with the old method of obtaining and evaluating the data